Sun set view by MailChimp
You are given an array representing the heights of neighboring buildings on a city street , from east to west. The city assessor would like you to write an algorithm that returns how many of these buildings have a view if the setting sun,in order to properly value to the street.
i/p : [3,7,8,3,6,1]
o/p : 3 //since the top three floors of the buildings with heights 8,6 and 1 all have an unobstructed view to the west.